requirements.txt
networkx==2.1
requests==2.18.4
setuptools==36.5.0.post20170921
tqdm==4.19.4
click==6.7
lxml==4.1.1
google_api_python_client==1.6.6
beautifulsoup4==4.6.0
PyYAML==3.12
